Make it look like a decent API
