+++ cupyx/distributed/__init__.py
@@ -1,4 +1,2 @@
-import cupy.cuda.nccl
-
-if cupy.cuda.nccl.available:
-    from cupyx.distributed._comm import NCCLBackend  # NOQA
+from cupyx.distributed._init import init_process_group  # NOQA
+from cupyx.distributed._comm import NCCLBackend  # NOQA
